+++ CHANGES.rst
@@ -28,6 +28,10 @@ Release history
   simulations don't run faster than real time.
   (`#85 <https://github.com/nengo/nengo-extras/pull/85>`_,
    `#151 <https://github.com/nengo/nengo/pull/151>`_)
+- Added ``nengo_extras.neurons.NumbaLIF``, which is a numba-accelerated
+  drop-in replacement for the ``nengo.LIF`` neuron model (requires ``numba`` to
+  be installed).
+  (`#86 <https://github.com/nengo/nengo-extras/pull/86>`_)
 
 0.3.0 (June 4, 2018)
 ====================